Rework CBA Settings component
