+++ addons/cba_settings/$PBOPREFIX$
@@ -1,0 +1,1 @@
+cba_settings_userconfig
