+++ addons/cba_settings/README.md
@@ -1,0 +1,10 @@
+# About
+
+CBA Settings Static Addon Userconfig. Does not conform to TAC Mods naming, made from [template](https://github.com/CBATeam/CBA_A3/tree/master/template/cba_settings_userconfig) provided by CBA.
+
+Only built correctly using `make.py`, `build.py` does not respect `$PBOPREFIX$`!
+
+### Authors
+
+- [CBA Team](http://github.com/CBATeam/CBA_A3)
+- [Jonpas](http://github.com/jonpas)
