--- addons/settings/$PBOPREFIX$
@@ -1,1 +1,0 @@
-x\tac\addons\settings
