--- addons/settings/CfgEventHandlers.hpp
@@ -1,5 +1,0 @@
-class Extended_PreInit_EventHandlers {
-    class ADDON {
-        init = QUOTE(call COMPILE_FILE(XEH_preInit));
-    };
-};
