--- addons/settings/config.cpp
@@ -1,17 +1,0 @@
-#include "script_component.hpp"
-
-class CfgPatches {
-    class ADDON {
-        name = COMPONENT_NAME;
-        units[] = {};
-        weapons[] = {};
-        requiredVersion = REQUIRED_VERSION;
-        requiredAddons[] = {"tac_main", "cba_settings"};
-        author = ECSTRING(main,Author);
-        authors[] = {"Jonpas"};
-        url = ECSTRING(main,URL);
-        VERSION_CONFIG;
-    };
-};
-
-#include "CfgEventHandlers.hpp"
